requirements.txt
flask==2.1.3
werkzeug==2.1.2
flask-sqlalchemy==2.5.1
flask-restx==0.5.1
python-dotenv==0.21.0
flask-wtf==1.0.1
email_validator==1.3.0
psycopg2==2.8.6
numpy